A seperate folder for all models of CU Fitness app is created. Also the models of myapp folder is transferred to this folder
